Fix build error on windows.
There is a redefined link error on windows platform, we need to
include "node.h" before requiring other headers.

Refer to https://groups.google.com/forum/#!topic/nodejs/onA0S01INtw

+++ src/greenworks_api.cc
@@ -98,7 +98,8 @@ NAN_METHOD(GetSteamId) {
   result->Set(NanNew("type"), GetSteamUserCountType(user_id.GetEAccountType()));
   result->Set(NanNew("accountId"), NanNew<v8::Integer>(user_id.GetAccountID()));
   result->Set(NanNew("staticAccountId"),
-              NanNew<v8::Integer>(user_id.GetStaticAccountKey()));
+              NanNew<v8::Integer>(static_cast<unsigned int>(
+                  user_id.GetStaticAccountKey())));
   result->Set(NanNew("isValid"), NanNew<v8::Integer>(user_id.IsValid()));
   result->Set(NanNew("level"), NanNew<v8::Integer>(
         SteamUser()->GetPlayerSteamLevel()));
